Separate http server to a new file
